# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Al Barakah Premium API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cronBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
...
    "/reviews": {
      "get": {
        "summary": "Public reviews",
        "security": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Wrong role"
          }
        }
      },
      "post": {
        "summary": "Create a review",
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "string"
                  },
                  "rating": {
                    "type": "number"
                  },
                  "comment": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  }
                },
                "required": [
                  "productId",
                  "rating",
                  "comment"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Wrong role"
          }
        }
      }
    },
...
    "/admin/reviews": {
      "get": {
        "summary": "Admin reviews",
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Wrong role"
          }
        }
      }
    },
...
    "/admin/backup": {
      "get": {
        "summary": "Download backup",
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Wrong role"
          }
        }
      }
    },
    "/admin/backup/restore": {
      "post": {
        "summary": "Restore backup",
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Wrong role"
          }
        }
      }
    },
...
  }
}